Add a new MakeParentDirectoryForPromise() function
We need a function for making parent directories for promised
files/directories/links/... that respects the eval mode (dry-run)
as well as the promise's action policy and that properly logs and
records failures.

The original MakeParentDirectory() function is just a wrapper
around the new function passing no promise information and
ignoring eval mode (because we need to create dictories in audit
mode). And we no longer need the MakeParentDirectory2() function
that only went half way to respecting eval mode and promise
action policy.

Ticket: ENT-5292
Changelog: None