Streamline logic in checking if link is already correct in VerifyLink()
Let's just use a proper 'bool' variable with a self-explanatory
name instead of 'int ok' which actually means the link is not
correct. Also, use an assert to make sure VerifyLink() is only
called when symbolic link is requested (there is the
VerifyHardlink() function for hard links) instead of complicating
the logic even further.

Ticket: ENT-5292
Changelog: None